--- packages/eslint-plugin-rax-runtime-miniapp/docs/rules/no-spread-props.md
@@ -1,35 +1,0 @@
-# 不推荐使用 {..props} 的方式传参 (no-spread-props)
-
-Please describe the origin of the rule here.
-
-## Rule Details
-
-This rule aims to...
-
-Examples of **incorrect** code for this rule:
-
-```js
-
-// fill me in
-
-```
-
-Examples of **correct** code for this rule:
-
-```js
-
-// fill me in
-
-```
-
-### Options
-
-If there are any options, describe them here. Otherwise, delete this section.
-
-## When Not To Use It
-
-Give a short description of when it would be appropriate to turn off this rule.
-
-## Further Reading
-
-If there are other links that describe the issue this rule addresses, please include them here in a bulleted list.
