+++ packages/eslint-plugin-rax-runtime-miniapp/package.json
@@ -1,6 +1,6 @@
 {
   "name": "eslint-plugin-rax-runtime-miniapp",
-  "version": "0.0.1-0",
+  "version": "1.0.0",
   "description": "rax 运行时 eslint 插件",
   "keywords": [
     "eslint",
@@ -8,7 +8,7 @@
     "eslint-plugin"
   ],
   "author": "chenrongyan",
-  "main": "lib/index.js",
+  "main": "src/index.js",
   "scripts": {
     "lint": "eslint .",
     "test": "mocha tests --recursive"
