--- packages/eslint-plugin-rax-runtime-miniapp/src/utils/props.js
@@ -1,2 +1,0 @@
-'use strict';
-
