--- packages/eslint-plugin-rax-runtime-miniapp/tests/src/rules/no-spread-props.js
@@ -1,34 +1,0 @@
-/**
- * @fileoverview 不推荐使用 {..props} 的方式传参
- * @author chenrongyan
- */
-"use strict";
-
-//------------------------------------------------------------------------------
-// Requirements
-//------------------------------------------------------------------------------
-
-const rule = require("../../../src/rules/no-inline-styles"),
-  RuleTester = require("eslint").RuleTester;
-
-const ruleTester = new RuleTester({
-  parserOptions: {
-    ecmaVersion: 10,
-    ecmaFeatures: {
-      jsx: true,
-    },
-  },
-});
-
-ruleTester.run("no-spread-props", rule, {
-  valid: [
-    // give me some code that won't trigger a warning
-  ],
-
-  invalid: [
-    {
-      code: "",
-      errors: [{ message: "Fill me in.", type: "Me too" }],
-    },
-  ],
-});
